LayoutTests/fast/js/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.2_Keywords/S7.5.2_A1.18-expected.txt

S7.5.2_A1.18

PASS ReferenceError: Left side of assignment is not a reference.

TEST COMPLETE
